% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/meta.R
\name{add_meta_prototype}
\alias{add_meta_prototype}
\title{Add 'prototype' attribute to data}
\usage{
add_meta_prototype(data)
}
\arguments{
\item{data}{a data object}
}
\description{
Add \code{prototype} attribute to data.  The \code{prototype} attribute is
a zero-row data.frame
}
\seealso{
\code{\link{meta_prototype}}
}

