% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{add_meta_values}
\alias{add_meta_values}
\title{Add 'values' attribute to data}
\usage{
add_meta_values(data)
}
\arguments{
\item{data}{a data object}
}
\description{
Add metadata 'values' attribute that containsa list of values associated
with categorical variables associated with a data frame
}
\seealso{
\code{\link{meta_values}}
}

