% Generated by roxygen2 (4.0.1): do not edit by hand
\name{apply.prototype}
\alias{apply.prototype}
\title{apply.prototype}
\usage{
apply.prototype(data, proto, default = "__OTHER__")
}
\arguments{
\item{data}{data}

\item{proto}{prototype}

\item{default}{character; value to be used when collapsing categories ( default: "__OTHER__" )}
}
\description{
Applies a prototype to data in a higly efficient way in order to
prepare the data for soc
}
\seealso{
\code{\link[formulations]{conform}}
  \code{\link[plyr]{rbind.fill}}
}

