% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.performance.R
\name{as.data.frame.performance}
\alias{as.data.frame.performance}
\title{Convert ROCR::performance object into a data.frame}
\usage{
\method{as.data.frame}{performance}(x, row.names = NULL, optional = FALSE,
  ...)
}
\arguments{
\item{x}{ROCR::performance object}

\item{row.names}{NULL or character; see \code{\link[base]{as.data.frame}}}

\item{optional}{logical; see \code{\link[base]{as.data.frame}}}

\item{...}{additional arguments passed to/from other methods}
}
\description{
Convert ROCR::performance object into a data.frame
}
\details{
Converts a performance object into a data.frame of performance measures useful 
for creating plots
}
