\docType{methods}
\name{assess}
\alias{assess}
\alias{assess,integer,integer-method}
\alias{assess,numeric,numeric-method}
\alias{assess,train,data.frame-method}
\title{Assess the fit between a value and its estimate}
\description{
  Assesses the fit between a value and its estimate using a
  variety of measures
}
\details{
  Calculates a variety of perfomance measures for a model
  fit.

  \code{train}, \code{data.frame} method evaluates the
  model
}

