% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calibrate}
\alias{calibrate}
\title{Calibrate}
\usage{
calibrate(x, y)
}
\arguments{
\item{x}{numeric; original distribution}

\item{y}{numeric; modeled distribution

Uses Sturges formula to...
Uses linear model to determine the calibrated ...}
}
\value{
numeric vector; y transformed into basis of x
}
\description{
Calculates one modelled distribution to closely match the original input distribution
}
\examples{
calibrate( 1:10, 1:10 ) # 1:10
   calibrate( 1:10, 1:20 )
   calibrate( 1:20, 1:10 )
   library(testthat)
   expect_is( x, 'numeric' )
}
\seealso{
reference any functions that are input links to other functions using latex \link{name function}
}

