% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confuse.R
\name{confuse}
\alias{confuse}
\alias{confuse.default}
\alias{confuse.categories}
\alias{confuse.character}
\alias{confuse.factor}
\alias{confuse.train}
\title{confuse}
\usage{
Method exists for caret::train() objects

\method{confuse}{default}(obs, pred, na.rm = TRUE)

\method{confuse}{categories}(obs, pred)

\method{confuse}{character}(...)

\method{confuse}{factor}(...)

\method{confuse}{train}(obs, pred = NULL)
}
\arguments{
\item{obs}{vector of observed/actual values}

\item{pred}{vector of predicted values}

\item{na.rm}{whether to removing missing values}

\item{...}{additional arguments

By convention, the \code{observed} values are placed in rows and the 
predicted values are placed in columns. The \code{dimnames} of the 
resulting matrix are consequently \code{observed} and \code{predicted}}
}
\value{
2x2 confusion matrix / contingency table
}
\description{
Create a confusion matrix for categorical values
}
\note{
- By convention the observed values are placed in rows, predicted 
  values are placed in columns
- observed
}
\examples{
  confuse.categories( obs=qw(A,A,B,B), pred=qw(A,B,C,C) )
  
  
}
\seealso{
\code{table} for cross-classifying factors
}
