% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/date-features.R
\name{featurize_date}
\alias{featurize_date}
\title{Create features from a date}
\usage{
featurize_date(dates, label = NULL, ...)
}
\arguments{
\item{dates}{date vector}

\item{label}{how to label the resulting features

Calculates various attributes for a given date.}

\item{...}{additional arguments}
}
\value{
a data.table with the following attributes: year, month, day,
week (week of the year), wday (day of the week), holiday (logical-
whether the day is a holiday, days_to_holiday, days_from_holiday
}
\description{
Create categorical date features
}
\examples{
featurize_date( now() )
}

