% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_model.R
\name{eval_model}
\alias{eval_model}
\alias{eval_model.default}
\title{Calculate model performance}
\usage{
eval_model(model, newdata, fun, ...)

\method{eval_model}{default}(model, newdata, fun = rmse, ...)
}
\arguments{
\item{model}{a model objecy. The model to be evaluated.}

\item{newdata}{Data on which to evaluate the model. Usually the test set.}

\item{fun}{A function that accepts (at least) two arguments the fitted and 
the actual value}

\item{...}{Arguments passed to \code{fun}}
}
\value{
The result of \code{fun} applied to the actual and fitted value
}
\description{
Calculates the performance of a fitted model.
}
\note{
Does not allow for post model evaluation
}
\examples{
  fit <- lm( Sepal.Length ~ ., iris )
  eval_model( fit, iris )
  eval_model( fit, iris, rmse )
  eval_model( fit, iris, mae )
  
}
\author{
Christopher Brown
}
\seealso{
\code{mae}, \code{rmse}
}

