% Generated by roxygen2 (4.0.2): do not edit by hand
\name{features}
\alias{features}
\title{features - a data.table of features}
\usage{
features(data, key = NULL)
}
\arguments{
\item{data}{to become features}

\item{key}{character; names of columns to use as the key, \code{NULL} the
default will use either existing keys or \code{ getOption('feature.key') }

The function works by converting the data to a data.table and attempting to
set a key on the data. The key will be determined by the explicit \code{key}
argument to this function, any existing key or the global option,
\code{feature.key}. If no key is defined a warning is issued.}
}
\value{
a data.table that includes class 'features' with an optional key set.
}
\description{
Create a S3 feature data.table by blessing into a class an optionally
setting the key for the data set
}
\examples{
features( iris, key='Species' )
  features( mtcars, "cyl" )
}

