% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generate}
\alias{generate}
\title{generate features}
\usage{
generate(x, data, weight = NULL, sep = ".", fun = sum)
}
\arguments{
\item{x}{(character) name of column to convert}

\item{data}{(data.frame or data.table)}

\item{weight}{numeric; optional weighting}

\item{sep}{character; character separator}

\item{fun}{function;}
}
\value{
a set of features ...
}
\description{
generate features
}
\details{
generates features based on levels of a categorical
variable.
}
\note{
- This is supposed to work by splitting according to
 - TODO: make this more like \code{\link{reshape}}
}
\examples{
generate( data=iris, x="Species" )
}
\seealso{
dummies
}

