% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getFeatures.r
\name{getFeatures}
\alias{getFeatures}
\title{getFeatures (DEPRECATED?)}
\usage{
getFeatures(data, by, features)
}
\arguments{
\item{data}{data}

\item{by}{by}

\item{features}{features}
}
\description{
Possibly deprecated function for evaluating a list of \code{features} from
\code{data}
}
\seealso{
\code{divide}
}

