% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggROC.R
\name{ggROC}
\alias{ggROC}
\alias{ggROC.numeric}
\alias{ggROC}
\alias{ggROC.train}
\alias{ggROC.caret}
\title{ggROC}
\usage{
ggROC(x, y, ...)

\method{ggROC}{numeric}(x, y, measure = "tpr", x.measure = "fpr", ...)

\method{ggROC}{train}(x, y, ...)
}
\arguments{
\item{x}{object}

\item{y}{depends on typeof(x). See details.}

\item{...}{additional arguments}

\item{measure}{string; measure used with \code{ROCR::performance}}

\item{x.measure}{string; measure used with \code{ROCR::performance}}
}
\description{
Produce a ROC curve using ggplot2
}
\details{
\code{ggROC} is an S3 generic function for producing ggplot based ROC curves.

For numeric objects, \code{x} is a predicted value, \code{y} is the labels.
}
\examples{

  data(ROCR.simple)
  ggROC( ROCR.simple$predictions, ROCR.simple$labels)
  perf <- performance(pred,"tpr","fpr")
  plot.performance(perf)

 ctrl <- trainControl(savePrediction=TRUE, classProbs = TRUE)
 fit <- caret::train( Species ~ ., iris, method="rf", metric="Kappa", trControl=ctrl )
 ggROC(fit)
 
}
\seealso{
\href{ROCR::plot.performance()}{ROCR::plot.performance}

\code{\link{plot.performance}}
}
