% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggSeparate.R
\name{ggSeparate}
\alias{ggSeparate}
\title{ggSeperate}
\usage{
ggSeparate(x, y)
}
\arguments{
\item{x}{object such as a vector}

\item{y}{object depends on x}
}
\description{
Separation plot for a numeric values and classes  
Separation
}
