% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_multinomial.R
\name{glm_multinom}
\alias{glm_multinom}
\title{Mutlinomial glm}
\usage{
glm_multinom(formula, data, ...)
}
\description{
Fits a multinomial logistic regression using glm models.
}
\examples{

 data(iris)
 fit <- glm_multinom( Species ~ . , data=iris)
 
   
}
