% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{ifeatures}
\alias{ifeatures}
\title{ifeatures - iterator over features}
\usage{
ifeatures(x, chunkSize = 1, ...)
}
\arguments{
\item{x}{data.table containing a key and measures}

\item{chunkSize}{number of measures to include along with

A feature is defined as a measurement associated with a particular entity.
In this case the entity is defined by the key of the data.table. ifeatures
iterators over the non-key columns}
}
\description{
iterator for iterating of a data.table of features
}
\examples{
data(iris)
  setDT(itis)
  setkey(iris, Species)

  it <- ifeatures(iris)
  while( hasNext(it) )
    message( paste( dim(nextElem(it)), collapse=" x " ) )
}
\seealso{
\code{\link{divide}}
  \code{\link[itertools]{isplitVector}}
  \code{\link[itertools]{ihasNext}}
}

