% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp.R
\name{imp}
\alias{imp}
\alias{importance}
\title{imp}
\usage{
imp(metric, ref, decreasing = FALSE)

importance(metric, ref, decreasing = FALSE)
}
\arguments{
\item{metric}{numeric; value of the metric}

\item{ref}{numeric; reference value of the metric, often the "naive" metric}

\item{decreasing}{logical; metrics improves by decreasing instead of 
increasing}
}
\description{
Report the improvement in a linear metric
}
\details{
Calculated as:

    sign * ( metric/ref - 1 )
}
\examples{

  metric <- 1.1 
  ref <- 1.0 
  
  imp( metric, ref )
  imp( metric, ref, decreasing = TRUE )
  
}
