% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute}
\alias{impute}
\alias{impute.POSIXct}
\alias{impute.character}
\alias{impute.data.frame}
\alias{impute.data.table}
\alias{impute.default}
\alias{impute.factor}
\alias{impute.integer}
\alias{impute.logical}
\alias{impute.numeric}
\title{Impute missing values}
\usage{
impute(x, fun, ...)

\method{impute}{default}(x, fun, ...)

\method{impute}{logical}(x, fun, ...)

\method{impute}{numeric}(x, fun = mean, ...)

\method{impute}{integer}(x, fun = median, ..., na.rm = TRUE)

\method{impute}{character}(x)

\method{impute}{factor}(x)

\method{impute}{POSIXct}(x, fun = mean, ..., na.rm = TRUE)

\method{impute}{data.frame}(x, fun, ...)

\method{impute}{data.table}(x, fun, ...)
}
\arguments{
\item{x}{object}

\item{fun}{function.}

\item{...}{other arguments passed to fun}

\item{na.rm}{boolean. Whether to remove \code{NA}s.  (Default:TRUE)

Imputes missing values through using a flexible function \code{fun}.

\code{impute.data.table} skips imputing columns where all value are \code{NA}}
}
\value{
An object with the same class as \code{x} with missing values replaced by
the most commonly occuring value(s), i.e. the mode for numeric values.  When
there is a tie for multiple values, the replacement is sampled from these.
}
\description{
Imputes missing values for an object, \code{x}
}
\examples{
x <- c( 1, NA, 3:5 )
impute( x, mean)   

x <- as.factor(mtcars$cyl )
x[1:4] <- NA

}

