% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_binomial_model.R
\name{is_binomial_model}
\alias{is_binomial_model}
\alias{is_binomial_model.caret}
\title{is_binomial_model}
\usage{
is_binomial_model(x)

\method{is_binomial_model}{caret}(x)
}
\description{
is_binomial_model
}
