% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/date-features.R
\name{isHoliday}
\alias{isHoliday}
\title{isHoliday}
\usage{
isHoliday(dates, tags = NULL)
}
\arguments{
\item{dates}{POXSIXct vector to test}

\item{tags}{character; one or more tags indication that

\code{isHoliday} determines whether dates fall on a holiday}
}
\value{
logical of the same length as \code{dates} indicating whether the date
  follows on a holiday
}
\description{
isHoliday
}
\examples{
isHoliday( ymd( 20140901, 20131225 ) )
  isHoliday( now() )
  isHoliday( ymd('20140901') )
}
\seealso{
\code{\link[TimeWarp]{isHoliday}}
}

