% Generated by roxygen2 (4.0.1): do not edit by hand
\name{logistic}
\alias{inv.logistic}
\alias{logistic}
\alias{logistic_inv}
\alias{logistic_inv,}
\alias{logistics,}
\title{logistic and inverse logistic function}
\usage{
logistic(x)

logistic_inv(y)
}
\arguments{
\item{x}{numeric. Argument passed to the logistic functions}

\item{y}{numeric. Argument passed to the inverse logistic function}
}
\description{
apply the logistic function to \code{x}
}

