% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mae}
\alias{mae}
\title{Compute the median absolute error (MAE) of an estimate.}
\usage{
mae(x, xhat)
}
\arguments{
\item{x}{numeric. The actual/true values.}

\item{xhat}{numeric. The estimated/predicted values.}
}
\value{
numeric. The calculated median absolute error.
}
\description{
Computes the median absolute error of en estimate or model.
}
\details{
The operation is commutative; the order of arguments does not matter. If
either \code{x} or \code{xhat} have missing values, then that tuple is
omitted from the calculations similar to \code{na.omit}.
}

