% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mase.R
\name{mase}
\alias{mase}
\title{Mean Absolute Scaled Error}
\usage{
mase(f, x, naive = naive, ...)
}
\arguments{
\item{f}{forecast;}

\item{naive}{function; forecast method used}

\item{...}{}

\item{actual}{numeric or time-series of observed response}
}
\description{
Mean Absolute Scaled Error
}
\details{
The mean absolute scaled error calculates the error relative to a naive
prediction
}
\references{
\url{https://www.otexts.org/fpp/2/5}
}
