% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/match_vars.R
\name{match_vars}
\alias{match_vars}
\title{Add dummy columns to variables data set}
\usage{
match_vars(data1, data2)
}
\arguments{
\item{data1}{data set to which columns are added.}

\item{data2}{data set to get the columns are taken.

\code{match_vars} adds columns to \code{data1} that are unmatched from
\code{data2}.  The columns added are given defaults \code{FALSE} for PA
variables, \code{0} for VAL variables and \code{0} for ERR variables.}
}
\value{
a data set like \code{data1} with unmatched columns of \code{data2}
added.  These columns are given deefault values.
}
\description{
Adds dummy columns to the first data frame that are present in the second
}
\seealso{
\code{\link{match_features}}
  \code{\link[plyr]{rbind.fill}}
}

