% Generated by roxygen2 (4.0.2): do not edit by hand
\name{measures}
\alias{measures}
\title{measures - a data.table of measures}
\usage{
measures(data, key = NULL, index = NULL)
}
\arguments{
\item{data}{to become measures}

\item{key}{character; names of columns to use as the key, \code{NULL} the
default will use either existing keys or \code{ getOption('feature.key') }}

\item{index}{The function works by converting the data to a data.table and attempting to
set a key on the data. The key will be determined by the explicit \code{key}
argument to this function, any existing key or the global option,
\code{feature.key}. If no key is defined a warning is issued.}
}
\value{
a data.table that includes class 'measures' with an optional key set.
}
\description{
Create a S3 feature data.table by blessing into a class an optionally
setting the key for the data set
}
\examples{
measures( iris, key='Species' )
  measures( mtcars, "cyl" )
}

