% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/meta.R
\name{meta_prototype}
\alias{meta_prototype}
\title{Create prototype for data}
\usage{
meta_prototype(data)
}
\arguments{
\item{data}{a data object}
}
\description{
Create a prototype for the data.  A prototype is a zero-row copy that retains
all attributes, columns, etc. just without data.
}
\seealso{
\code{\link{add_meta_prototype}}
}

