% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{meta_values}
\alias{meta_values}
\title{Create a list of values associated with data}
\usage{
meta_values(data)
}
\arguments{
\item{data}{a data object}
}
\description{
Creates a list based on a data. The list contains the values found in the 
categorical variables of the data. The
}
\seealso{
\code{\link{add_meta_values}}
}

