% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parRandomForest.R
\name{parRandomForest}
\alias{parRandomForest}
\alias{randomForest}
\title{parRandomForest}
\usage{
parRandomForest(..., ntree = 500)

randomForest(...)
}
\arguments{
\item{...}{parameters passed to \code{randomForest}}

\item{ntree}{integer; number of trees train

Trains a Random Forest model using a \code{foreach} cluster. The cluster must
first be registered with \code{registerDoParallel} 

If \code{getOption('verbose')} is TRUE, a message is printed. This helps to
check usage in package like 'caret'.}
}
\value{
A \code{randomForest} object.
}
\description{
provides parRandomForest and a parallel drop-in replacement for 
\code{randomForest}using foreach/doParallel
}
\examples{
  registerDoParallel(4)
  
  #EXPLICIT:
  fit <- parRandomForest( Species ~ . , iris, ntree=200)
  
  # DROP-IN REPLACEMENT
  fit <- randomForest( Species ~ . , iris, ntree=200)
  
  # CARET 
  fit <- train( Species ~ . , iris, ntree=200 )
  
}
\references{
\url{http://stackoverflow.com/questions/5571774/what-is-the-easiest-way-to-parallelize-a-vectorized-function-in-r/15414748#15414748}
}
\seealso{
\code{\link[randomForest]{randomForest}}, 
  code{\link[caret]{train}}, 
  \code{\link[doParallel]{registerDoParallel}},
  \code{\link[foreach]{foreach}}
}

