% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.performance.R
\docType{methods}
\name{plot,performance,missing-method}
\alias{plot,performance,missing-method}
\title{plot.performance}
\usage{
\S4method{plot}{performance,missing}(x, y)
}
\arguments{
\item{x}{Object to plot}

\item{y}{Missing}
}
\value{
a ggplot2 object
}
\description{
A drop-in replacement for ROCR::plot.perforamnce that uses ggplot.
}
\examples{
  library(ROCR)
  data(ROCR.simple)
  pred <- prediction( ROCR.simple$predictions, ROCR.simple$labels)
  perf <- performance(pred,"tpr","fpr")
  plot(perf)

}
\seealso{
\code{\link[ROCR]{plot.performance}}
  \code{\link{as.data.frame.performance}}
}
\author{
Decision Patterns
}
