% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.performance.R
\name{plot.performance}
\alias{plot.performance}
\title{plot.performance}
\usage{
plot.performance(x, y)
}
\value{
a ggplot2 object
}
\description{
A drop-in replacement for ROCR::plot.perforamnce that uses ggplot.
}
\examples{

qplot( perf@x.values[[1]], perf@y.values[[1]])

  library(ROCR)
  data(ROCR.simple)
  pred <- prediction( ROCR.simple$predictions, ROCR.simple$labels)
  perf <- performance(pred,"tpr","fpr")
  plot.performance(perf)

}
\seealso{
\code{\link[ROCR]{plot.performacne}}
}
\author{
Decision Patterns
}
