% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.univariates}
\alias{plot.univariates}
\title{plot.univariates}
\usage{
\method{plot}{univariates}(...)
}
\arguments{
\item{...}{one or more vectors

\code{plot.univariates} is useful for comparing two or more distributions}
}
\value{
a ggplot object having one facet and color for each distribution
}
\description{
Create a ggplot of one or more distributions as a comparison
}
\examples{
plot.univariates( norm=rnorm(1000, 5), lnorm=rlnorm(10000, 10) ) + scale_x_sqrt()
 plot.univariates( 1:100, rlnorm=rlnorm(10000, 10) ) + scale_x_sqrt()
 plot.univariates( 1:100, rnorm(100))
}

