% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_univariates.R
\name{plot_univariates}
\alias{plot_univariates}
\title{plot_univariates}
\usage{
plot_univariates(...)
}
\arguments{
\item{...}{one or more vectors

\code{plot_univariates} is useful for comparing two or more distributions}
}
\value{
a ggplot object having one facet and color for each distribution
}
\description{
Create a plot of one or more distributions as a comparison
}
\examples{
 \dontrun{ 
   plot_univariates( norm=rnorm(1000, 5), lnorm=rlnorm(10000, 10) ) + scale_x_sqrt()
   plot_univariates( 1:100, rlnorm=rlnorm(10000, 10) ) + scale_x_sqrt()
   plot_univariates( 1:100, rnorm(100))
 }
}

