% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_multinomial.R
\name{predict.glm_multinom}
\alias{predict.glm_multinom}
\title{Predict multinomial GLM model}
\usage{
\method{predict}{glm_multinom}(x, newdata, type = c("norm", "probs", "raw",
  "class"), ...)
}
\arguments{
\item{x}{glm_multinomial object}

\item{newdata, }{data.frame}

\item{type}{string; one of 'norm' (default), 'probs', 'raw' or 'class'}

\item{...}{unused}
}
\value{
- For `type="raw"`  : A matrix with nrow(newdata) rows and one prediction per model
 - For `type="probs"`: A matrix of raw probabilities
 - For  type="norm"  : A matrix of normalized probabilities 
 - For  type="class" : A vector of predicted classes
}
\description{
Predicts a multinomial GLM model
}
\details{

}
