% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prep_data}
\alias{prep_data}
\title{Prepare data for use by a particular function}
\usage{
prep_data(data, fun, ...)
}
\arguments{
\item{data}{data.frame.}

\item{fun}{name of function to prep data for}

\item{...}{arguments passed to other functions}
}
\value{
a data frame.
}
\description{
Prepares a data set for use by a particular function
}
\examples{
prep_data( iris, randomForest )
}

