% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data_randomForest.R
\name{prep_data_randomForest}
\alias{prep_data_randomForest}
\title{Prepare Data for randomForest}
\usage{
prep_data_randomForest(data, ..., max.levels = 50, preserve = character(),
  keep = "Other")
}
\arguments{
\item{data}{data.frame}

\item{...}{arguments passed to other functions}

\item{max.levels}{integer; maximum number of levels for categorical variables}

\item{preserve}{(character) columns not to be prepared}
}
\description{
Prepare Data for randomForest
}
\details{
The Random Forest package requires: 
  factors and not strings, 
  factors <= 50 levels, 
  no missing values,
  no +/- Inf values
   
Additionally, it requires that subsequent factors not have newlevels in the 
scoring data.  Each level does not have to be represented, but there needs
all levels must be represent in the \code{levels} attribute.
   
also saves 'values',
}
\examples{
  prep_data_randomForest( iris )
  
}

