% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prep_data_scoring.R
\name{prep_score}
\alias{prep_score}
\title{Prepare data for training and scoring of a caret model}
\usage{
prep_score(data, model, OTHER = "__OTHER__")
}
\arguments{
\item{data}{data.table to prepare}

\item{model}{train caret model to emulate}

\item{OTHER}{character; value to be used when there are too many categorical
levelss}
}
\value{
a data.table that can (hopefully) be used by the \code{predict}
method
}
\description{
Prepares data for scoring of a caret model by making sure scoring data is
similar to trainingData.
}
\details{
Prepares data for scoring does the following
 - checks for missing trainingData
 - checks missing columns in newdata (n.b. this would be caught by predict)
 - reduce new levels to OTHER value.
 - drop inused levels
 - Impute missing values
}
\note{
OTHER could be \code{NA}, but there is a difference between missing
and present but reduced. This maintains
}
\seealso{
prep_data, reduce_cardinality
}

