% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/prep_data_scoring.R
\name{prep_train}
\alias{prep_train}
\title{randomForest requires
* factors and not strings.
* factors <= 32 levels
* no missing y values
* also saves 'values'
* Inf, -Inf not allowed (at least not in caret )
* predict additionally requires that there are no new levels in
  \code{newdata}. Each level does not have to be represented, but there needs
  all levels must be represent in the \code{levels} attribute.}
\usage{
prep_train(data, ..., preserve = character())
}
\arguments{
\item{data}{data.frame.}

\item{...}{arguments passed to other functions}

\item{preserve}{(character) columns not to prepare.}
}
\description{
randomForest requires
* factors and not strings.
* factors <= 32 levels
* no missing y values
* also saves 'values'
* Inf, -Inf not allowed (at least not in caret )
* predict additionally requires that there are no new levels in
  \code{newdata}. Each level does not have to be represented, but there needs
  all levels must be represent in the \code{levels} attribute.
}
\examples{
# -tk
}
\seealso{
\code{\link{prep_data}}
}

