% Generated by roxygen2 (4.0.2): do not edit by hand
\name{probs}
\alias{probs}
\alias{probs.randomForest}
\title{probs}
\usage{
probs(object, newdata, ...)

\method{probs}{randomForest}(object, newdata, correct = FALSE, ...)
}
\arguments{
\item{object}{a model;}

\item{newdata}{data; data for which to calculate probaiities}

\item{...}{arguments passed to \code{predict.randomForest}

\code{probs} is an S3 generic method.  Functions can be written for
classification models.

Parameter \code{correct} determines whether the final probabilities are
adjusted based on the frequency of occurence in the training data. This is
useful when the training set does not represent the proportion of
observations, but rather is used to define the boundaries of that space. This
is important when using the Brieman column scramble technique.}

\item{correct}{logical; whether to correct the probabilities according to their
frequency of occurance.}
}
\value{
numeric matrix of prediction probabilities with the same number of
  rows as \code{newdata} and as many columns as classes in the fitted model
  \code{object}
}
\description{
Get matrix of predicted class probabilities for a model optionally
correcting for imbalances in classes in the training sample.
}
\note{
Although not implemented here, the use \code{probs} could be used to make a
 Bayesian type inference if the frequency of occurence of the final outcome
 is known.

 A better method would probably be to adjust weights during model training
}
\examples{
f <- randomForest( Species ~ ., iris[1:125,] )
  probs( f, iris )
  probs( f, iris, correc=TRUE )
}
\seealso{
\code{\link[randomForest]{predict.randomForest}}
}

