% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/r2.R
\name{r2}
\alias{r2}
\title{Compute pseudo-rsquared for a model}
\usage{
r2(x, xhat)
}
\arguments{
\item{x}{the TRUE amount}

\item{xhat}{estimated amount}
}
\description{
Computes the psuedo-rsquared for a model
}

