% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{rmse}
\alias{rmse}
\title{Compute the root mean squared error}
\usage{
rmse(x, xhat)
}
\arguments{
\item{x}{numeric. The true/actual values.}

\item{xhat}{numeric. The predicted/estimated values.}
}
\description{
Computes the root mean square error (rmse) of en estimate or model.
}
\details{
The operation is commutative so the order of operations does not matter. If 
either \code{x} or \code{xhat} have missing values, then that tuple is 
omitted from the calculations similar to \code{na.omit}.
}

