% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scramble.R
\name{scramble}
\alias{scramble}
\alias{scramble.data.frame}
\alias{scramble.matrix}
\title{Column scramble}
\usage{
scramble(x)

\method{scramble}{data.frame}(x)

\method{scramble}{matrix}(x)
}
\arguments{
\item{x}{Object to scramble the columns of

Scrambles the columns of a matrix or data.frame. This is useful for
creating supervised clustering models where there are no examples of
FALSE/not-present examples of the data. This is after Breiman.}
}
\value{
An object with the same class and size of \code{x} in which each
column has been randomly reordered.
}
\description{
Scrambles the columns of a matrix or data.frame
}
\note{
Incorporated with permission from the copyright holder.
}
\examples{
scramble(iris)
}
\author{
Christopher Brown
}
\references{
Brieman's Random Forest paper.
}

