% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softmax.R
\name{softmax}
\alias{softmax}
\title{softmax}
\usage{
softmax(x)
}
\arguments{
\item{x}{}
}
\description{
calculate the softmax of values
}
\examples{

x <- c(1, 2, 3, 4, 1, 2, 3)
softmax(x)

}
\seealso{
- [logistic](logistic())
}
