% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stitch.R
\name{stitch}
\alias{.stitch.outer}
\alias{.stitch.straight}
\alias{stitch}
\title{stitch two feature data tables together}
\usage{
stitch(x, y, ...)

.stitch.outer(x, y)

.stitch.straight(x, y, ...)
}
\arguments{
\item{x}{data.table;}

\item{y}{data.table;}

\item{...}{additional arguments passed to subsequent methods

\code{stitch} does two things.  First, it ensures removes duplicates from
\code{y}. This is done using \code{dup.action}.

merges (joins) \code{x} to \code{y} using a LEFT INNER JOIN. Special
behaviours make it so this according to indexes/columns of \code{x} and \code{y}.

If there are more than one \code{y} for every x. Duplicate rows of \code{y}
are pivoted until a one-to-one or many-to-one relationship exists, allowing
for a STRAIGHT JOIN.

If all indices of \code{y} are found in \code{x}, \code{y} are DEDUPED and
merged via a STRIAGHT JOIN to \code{x} via the indices of \code{y}.

If there are more

DEDUPLICATIONS are handled via the DEDUP action which defaults to attr( )
be additional attributes for x

\strong{imputation} is}
}
\value{
data.table; X and Y appropriately merged/joined.
}
\description{
stitches two data.tables together based on (shared) keys

\code{stitch.outer} is similar to/identical .stitch.straight

\code{.stitch.straight} performs a LEFT merge/join on two data.tables when \code{x} and \code{y} have
the key of \code{y} are all in the x
}
\note{
Indexes may not be necessary and this may work by matching names between
  \code{x} and \code{y}.

  LHS := Y's %intersect% X's
  RHS := Y's \ X's
}
\examples{
x <- data.table( customer=letters[1:4] )
  setkey( x, customer )

  y <- data.table(
       customer = sort( letters[ c(1:4,1:4,1:4) ] )
     , time     = 1:3
     , income   = round(rnorm(12,10) * 10)
     , expense  = -round( rnorm(12,10) )
     , count    = 1
  )

  y <- rbind(y, y[c(1,1,2),] )
  y <- y[ -(10:11), ]
  y[ 8, income := NA]
  setkey( y, customer, time )

  ml.tools:::.stitch.outer(x,y)
x <- data.table(
       customer = sort( letters[ c(1:4,1:4,1:4) ] )
     , date     = 1:3
     , income   = round(rnorm(12,10) * 10)
     , expense  = -round( rnorm(12,10) )
     , count    = 1
  )

  # DUPLICATES
  y <- data.table( customer=letters[1:2], first_name=c('Foo', 'Bar') )
  setkey( y, customer )

  ml.tools:::.stitch.straight( x, y  )
}
\seealso{
\code{\link[data.table]{merge}}
  \code{dup.pivot} in the \code{dup.actions} package.
}

