% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_roc.R
\name{style_roc}
\alias{style_roc}
\title{style_roc}
\usage{
style_roc(major.breaks = c(0, 0.1, 0.25, 0.5, 0.75, 0.9, 1),
  minor.breaks = c(seq(0, 0.1, by = 0.01), seq(0.9, 1, by = 0.01)),
  guide = TRUE, xlab = "False positive fraction",
  ylab = "True positive fraction", theme = theme_hc)
}
\description{
style_roc
}
\references{
plotROC
}
