% Generated by roxygen2 (4.0.1): do not edit by hand
\name{window_expr}
\alias{window_expr}
\title{window_expr}
\usage{
window_expr(milestone, event_dt, direction = "of", span)
}
\arguments{
\item{milestone}{character; name of milestone_dt column}

\item{event_dt}{character; name of date column related to the event}

\item{direction;}{before, after, within}

\item{span}{length of time for the window

Window := direction + span

before: Milestone - event_dt <= window  (LEADING INDICATOR)
after : Milestone - event_dt >= window  (LAGGING INDICATOR)
within (of): abs(Milestone - event_dt ) <= windows}
}
\value{
An unevalutated call object that can be evaluated in the milestone-event dataset.
}
\description{
generate an expression that can be \code{eval}'d on the event
data set to keep only the window of relevant events to the features
}
\examples{
window_expr( "milestone_dt", "event_dt", ddays(2) )
  window_expr( "milestone_dt", "event_dt", ddays(3), "within" )
  window_expr( "milestone_dt", "event_dt", ddays(4), "before" )
  window_expr( "milestone_dt", "event_dt", ddays(5), "after" )
}

