\name{apply.prototype}
\alias{apply.prototype}
\title{Applies a prototype to data in a higly efficient way in order to
prepare the data for soc}
\usage{
apply.prototype(data, proto, default = "__OTHER__")
}
\description{
Applies a prototype to data in a higly efficient way in
order to prepare the data for soc
}

