% Generated by roxygen2 (4.0.1): do not edit by hand
\name{apply.prototype}
\alias{apply.prototype}
\title{Applies a prototype to data in a higly efficient way in order to
prepare the data for soc}
\usage{
apply.prototype(data, proto, default = "__OTHER__")
}
\description{
Applies a prototype to data in a higly efficient way in order to
prepare the data for soc
}
\seealso{
\code{\link[formulations]{conform}}
  \code{\link[plyr]{rbind.fill}}
}

