% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{assess}
\alias{assess}
\alias{assess,integer,integer-method}
\alias{assess,numeric,numeric-method}
\alias{assess,train,data.frame-method}
\title{Assess the fit between a value and its estimate}
\usage{
assess(x, y, ...)

\S4method{assess}{integer,integer}(x, y, na.action = na.exclude, ...)

\S4method{assess}{numeric,numeric}(x, y, na.action = na.exclude,
  format = scales::comma, ...)

\S4method{assess}{train,data.frame}(x, y, na.action = na.exclude, ...)
}
\description{
Assesses the fit between a value and its estimate using a variety of measures
}
\details{
Calculates a variety of perfomance measures for a model fit.

\code{train}, \code{data.frame} method evaluates the model
}

