% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/assess.R
\docType{methods}
\name{assess}
\alias{assess}
\alias{assess,integer,integer-method}
\alias{assess,numeric,numeric-method}
\alias{assess,train,data.frame-method}
\title{Assess the fit between a value and its estimate}
\usage{
assess(x, y, ...)

\S4method{assess}{integer,integer}(x, y, na.action = na.exclude, ...)

\S4method{assess}{numeric,numeric}(x, y, na.action = na.exclude,
  format = scales::comma, ...)

\S4method{assess}{train,data.frame}(x, y, na.action = na.exclude, ...)
}
\arguments{
\item{x}{vector of outcomes}

\item{y}{vector of predictions}

\item{...}{additional arguments}

\item{na.action}{function to be used when arguments are absent}

\item{format}{(See scales package.)

Calculates a variety of perfomance measures for a model fit.}
}
\description{
Assesses the fit between a value and its estimate using a variety of measures
}
\details{
\code{train}, \code{data.frame} method evaluates the model
}

