% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confuse.R
\name{confuse}
\alias{confuse}
\alias{confuse.categories}
\alias{confuse.character}
\alias{confuse.default}
\alias{confuse.factor}
\title{2x2 confusion matrix for logical vectors}
\usage{
confuse(actual, predicted)

\method{confuse}{default}(actual, predicted, na.rm = TRUE)

\method{confuse}{categories}(actual, predicted)

\method{confuse}{character}(...)

\method{confuse}{factor}(...)
}
\arguments{
\item{actual}{logical vector of actual/observed values}

\item{predicted}{logical vector of predicted values}

\item{...}{additional arguments

By convention, the \code{actual} values are placed in rows and the 
predicted values are placed in columns. The \code{dimnames} of the 
resulting matrix are consequently \code{actual} and \code{predicted}}
}
\value{
2x2 confusion matrix / contingency table
}
\description{
2x2 confusion matrix for logical vectors
}
\note{
- By convention the actual values are placed in rows, predicted 
  values are placed in columns
- actual
}
\seealso{
\code{table} for cross-classifying factors
}

