% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{confuse}
\alias{confuse}
\alias{confuse.categories}
\alias{confuse.character}
\alias{confuse.default}
\alias{confuse.factor}
\title{2x2 confusion matrix for logical vectors}
\usage{
confuse(actual, predicted)

\method{confuse}{default}(actual, predicted)

\method{confuse}{categories}(actual, predicted)

\method{confuse}{character}(...)

\method{confuse}{factor}(...)
}
\arguments{
\item{actual}{logical vector of actual/observed values}

\item{predicted}{logical vector of predicted values}

\item{...}{additional arguments}
}
\value{
2x2 confusion matrix / contingency table
}
\description{
2x2 confusion matrix for logical vectors
}
\seealso{
\code{table} for cross-classifying factors
}

