% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{days_to_we_end}
\alias{days_to_we_end}
\title{Count days to the weekend's end}
\usage{
days_to_we_end(dates, use.holiday = TRUE, ...)
}
\arguments{
\item{dates}{Date objects}

\item{use.holiday}{Whether to add holidays are the end of
the weekend.}

\item{...}{additional arguments passed to \code{isHoliday} such as \code{tags}}
}
\value{
the length of time in days between the date and the end of the
weekend, including long weekends
}
\description{
Count the number of days from the date to the end of
the week, optionally accounting for Monday holidays
}
\details{
Considers Monday and Tuesday holidays, but will not consider
Wednesday ...
}
\examples{
days_to_we_end( now() )
 days_to_we_end( ymd(20140831) )
}

