% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/first.R
\name{first}
\alias{first}
\alias{first.data.table}
\alias{first.default}
\alias{nth}
\title{Get the first, last or nth record of a set of records}
\usage{
first(.data, group_by, arrange = NULL)

\method{first}{default}(.data, group_by = NULL, arrange = NULL)

\method{first}{data.table}(dt, key, o)

nth(.data, nth, group_by = NULL, arrange = NULL)
}
\arguments{
\item{.data}{A tbl.}

\item{group_by}{character; columns thay}

\item{arrange}{character; ordering columns}
}
\description{
Get the first, last or nth record of a set of records
}
\examples{
library(nycflights13)

  flights  \%>\% first( 'tailnum', c('year','month','day', 'dep_time') )
  flights  \%>\% last( 'tailnum', c('year','month','day', 'dep_time') )
  flights  \%>\% nth( 2, 'tailnum', c('year','month','day', 'dep_time') )

  first.data.table( flights, 'tailnum', c('year','month','day','dep_time') )
first.data.table( flights, 'tailnum', c('year','month','day','dep_time') )
}

