% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getImp}
\alias{getImp}
\title{getImp}
\usage{
getImp(model, n = length(model$importance))
}
\arguments{
\item{model}{A rf model or at least something that has an \code{importance}}

\item{n}{number of features to extract; the default is to extract all features}
}
\value{
factor
}
\description{
Get most important features from an RF model in order of importance
}
\examples{
# -tk
}
\seealso{
\code{\link[randomForest]{randomForest}}
}

