% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ifeatures}
\alias{ifeatures}
\title{ifeatures - iterator over features}
\usage{
ifeatures(x, chunkSize = 1, ...)
}
\arguments{
\item{x}{data.table containing a key and measures}

\item{chunkSize}{number of measures to include along with}

\item{...}{additional parameters passed to \code{ihasNext}

A feature is defined as a measurement associated with a particular entity.
The entities are defined by the key columns.  The measurements are non-key
columns.

nextElem iterates over the measurement (non-key) column and provides a
data.table with the keys of x and one measurement.

The idea is that the measurements can be independently operated on.}
}
\value{
an iterator
}
\description{
iterator for iterating of a data.table of features
}
\examples{
data(iris)
  setDT(iris)
  setkey(iris, Species)

  it <- ifeatures(iris)
  el1 <- nextElem(it)
  key(el1)
  while( hasNext(it) )
    message( paste( dim(nextElem(it)), collapse=" x " ) )
}
\seealso{
\code{\link{divide}}
  \code{\link[itertools]{isplitVector}}
  \code{\link[itertools]{ihasNext}}
}

