% Generated by roxygen2 (4.0.2): do not edit by hand
\name{isHoliday}
\alias{isHoliday}
\title{isHoliday}
\usage{
isHoliday(dates, tags = NULL)
}
\arguments{
\item{dates}{POXSIXct vector to test}

\item{tags}{character; one or more tags indication that

\code{isHoliday} determines whether dates fall on a holiday}
}
\value{
logical of the same length as \code{dates} indicating whether the date
  follows on a holiday
}
\description{
isHoliday
}
\examples{
isHoliday( ymd( 20140901, 20131225 ) )
  isHoliday( now() )
  isHoliday( '20140901' )
}
\seealso{
\code{\link[TimeWarp]{isHoliday}}
}

