% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ml.tools-package.R
\docType{package}
\name{ml.tools}
\alias{ml.tools}
\alias{ml.tools-package}
\title{ml.tools - Machine Learning Tools}
\description{
The \code{ml.tools} packages provides a number of tools that are generally
applicable for Machine learning applications.
}
\examples{
# -tk
}
\author{
Christopher Brown
}

