% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{nodes}
\alias{nodes}
\title{nodes}
\usage{
nodes(object, newdata)
}
\arguments{
\item{object}{Object to get the nodes from}

\item{newdata}{data.frame of data}
}
\value{
numeric matrix containing one row per observation, one column per tree in
  the forest. The values indicate which node of each tree the observation
  falls in.
}
\description{
Report the nodes numbers
}
\examples{
fit <- randomForest( Species ~ . , iris )
  nodes( fit, iris )
}
\seealso{
\code{\link{predict}}
}

