% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prep_data_randomForest}
\alias{prep_data_randomForest}
\title{Prepare Data for randomForest}
\usage{
prep_data_randomForest(data, ..., max.levels = 32, preserve = character())
}
\arguments{
\item{data}{data.frame.}

\item{preserve}{(character) columns not to be prepared}

\item{...}{arguments passed to other functions}
}
\description{
Random Forest requires:
* factors and not strings
* factors <= 32 levels
* no missing values
* also saves 'values'
* Inf, -Inf not allowed (at least not in caret )
* predict additionally requires that there are no new levels in
  \code{newdata}. Each level does not have to be represented, but there needs
  all levels must be represent in the \code{levels} attribute.
}
\examples{
# -tk
}

