% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/prep_data_randomForest.R
\name{prep_data_randomForest}
\alias{prep_data_randomForest}
\title{Prepare Data for randomForest}
\usage{
prep_data_randomForest(data, ..., max.levels = 50, preserve = character())
}
\arguments{
\item{data}{data.frame}

\item{...}{arguments passed to other functions}

\item{max.levels}{integer; maximum number of levels for categorical variables}

\item{preserve}{(character) columns not to be prepared}
}
\description{
Random Forest requires:
  factors and not strings,
  factors <= 32 levels,
  no missing values,
  also saves 'values',
  Inf, -Inf not allowed (at least not in caret )
}
\details{
predict additionally requires that there are no new levels in
  \code{newdata}. Each level does not have to be represented, but there needs
  all levels must be represent in the \code{levels} attribute.
}
\examples{
prep_data_randomForest( iris )
}

