% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/size.R
\name{size}
\alias{size}
\title{size}
\usage{
size(object)
}
\arguments{
\item{object,}{R object}
}
\description{
Show the size of the object (side-effect only)
}
\note{
TODO:
  - make generic
  - move into bigO(?) package
}
\examples{
size(iris)
  size(mtcars)
}

